Revert "fast rgf (#73)"
This reverts commit 2dad8ebbc89186f4a5f6e3dca3a1f66098120ea7.
