--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "include/fast_rgf"]
-	path = include/fast_rgf
-	url = https://github.com/baidu/fast_rgf
