+++ .travis.yml
@@ -13,25 +13,16 @@ env:
   - PYTHON_VERSION=3.5
   - PYTHON_VERSION=3.6
 
-git:
-  submodules: true
-
 before_install:
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
-      brew install gcc;
-      brew link --overwrite gcc;
-      export CXX=g++-7 && export CC=gcc-7;
       curl https://repo.continuum.io/miniconda/Miniconda${PYTHON_VERSION:0:1}-latest-MacOSX-x86_64.sh -o miniconda.sh;
     else
-      sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test;
-      sudo apt-get update;
-      sudo apt-get install g++-5;
-      export CXX=g++-5 && export CC=gcc-5;
       wget https://repo.continuum.io/miniconda/Miniconda${PYTHON_VERSION:0:1}-latest-Linux-x86_64.sh -O miniconda.sh;
     fi
 
   - bash miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
+  - hash -r
   - conda config --set always_yes yes --set changeps1 no
   - conda update -q conda
 
@@ -42,23 +33,14 @@ before_install:
 
 install:
   - conda install -q numpy scipy nose pandas scikit-learn
-  - pip install pytest>=3.3
   - python setup.py sdist --formats gztar
   - pip install dist/rgf_python-$RGF_VER.tar.gz -v
 
-# Temp solution for testing FastRGF
-  - cd include/fast_rgf/build
-  - cmake ..
-  - make
-  - make install
-  - cp ../bin/forest_train ../bin/forest_predict ~/
-  - cd ../../..
-
 script:
-  - pytest tests/ -o "python_files=*.py" -v
+  - python tests/test.py
   - pip uninstall -y rgf_python
   - python setup.py install
-  - pytest tests/ -o "python_files=*.py" -v
+  - python tests/test.py
 
 before_deploy:
   - if [[ $TRAVIS_OS_NAME == "osx" ]]; then
