+++ Readme.rst
@@ -14,8 +14,6 @@ Features
 
 Original RGF implementation is available only for regression and binary classification, but rgf\_python is also available for **multiclass classification** by "One-vs-Rest" method.
 
-**Alpha version** Fast RGF is supported. Please see `this guide <https://github.com/fukatani/rgf_python/blob/master/FastRGF.rst>`__
-
 Example:
 
 .. code:: python
@@ -226,7 +224,7 @@ rgf_python is distributed under the GNU General Public License v3 (GPLv3). Pleas
 
 rgf_python includes RGF version 1.2 which is distributed under the GPLv3. Original CLI implementation of RGF you can download at http://tongzhang-ml.org/software/rgf.
 
-Many thanks to Rie Johnson and Tong Zhang (the authors of RGF).
+We thank Rie Johnson and Tong Zhang (authors of RGF).
 
 Other
 -----
